✨ [feat] #51 영상 녹화 중 기울기 데이터 수집 기능 구현 추가
- 영상 녹화 중 1/3초마다 기울기 데이터를 수집하는 타이머를 CameraViewModel에 도입
- 타임스탬프가 포함된 기울기 값을 저장하도록 구현
- TimeStampedTilt의 시간 값을 Double 타입으로 변경
- 기울기 관측에 대한 Cancellable 배열에서 객체로 수정
- TiltDataCollector가 해제될 때 motion updates를 자동으로 중지하도록 처리

+++ Chalkak/Data/Model/Tilt.swift
@@ -27,7 +27,7 @@ struct Tilt: Codable, Hashable {
 /// 특정 시점의 기울기 정보를 담고 있는 구조체입니다.
 struct TimeStampedTilt: Codable {
     /// 시간 정보 (밀리초 단위).
-    let time: Int64
+    let time: Double
     
     /// 해당 시점의 카메라 기울기 값.
     let tilt: Tilt
